LayoutTests/imported/w3c/web-platform-tests/screen-wake-lock/idlharness.https.window-expected.txt


FAIL idl_test setup promise_test: Unhandled rejection with value: object "Error: unimplemented"
PASS idl_test validation
PASS Partial interface Navigator: original interface defined
PASS Partial interface Navigator: member names are unique
PASS Partial interface mixin NavigatorID: member names are unique
PASS Navigator includes NavigatorID: member names are unique
PASS Navigator includes NavigatorLanguage: member names are unique
PASS Navigator includes NavigatorOnLine: member names are unique
PASS Navigator includes NavigatorContentUtils: member names are unique
PASS Navigator includes NavigatorCookies: member names are unique
PASS Navigator includes NavigatorPlugins: member names are unique
PASS Navigator includes NavigatorConcurrentHardware: member names are unique
FAIL WakeLock interface: existence and properties of interface object assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock interface object length assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock interface object name assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock interface: existence and properties of interface prototype object assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock interface: operation request(optional WakeLockType) assert_own_property: self does not have own property "WakeLock" expected property "WakeLock" missing
FAIL WakeLock must be primary interface of navigator.wakeLock assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of navigator.wakeLock assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL WakeLock interface: navigator.wakeLock must inherit property "request(optional WakeLockType)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL WakeLock interface: calling request(optional WakeLockType) on navigator.wakeLock with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL WakeLockSentinel interface: existence and properties of interface object assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface object length assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface object name assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: existence and properties of interface prototype object assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: attribute released assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: attribute type assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: operation release() assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel interface: attribute onrelease assert_own_property: self does not have own property "WakeLockSentinel" expected property "WakeLockSentinel" missing
FAIL WakeLockSentinel must be primary interface of sentinel assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: sentinel"
FAIL Stringification of sentinel assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: sentinel"
FAIL WakeLockSentinel interface: sentinel must inherit property "released" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: sentinel"
FAIL WakeLockSentinel interface: sentinel must inherit property "type" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: sentinel"
FAIL WakeLockSentinel interface: sentinel must inherit property "release()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: sentinel"
FAIL WakeLockSentinel interface: sentinel must inherit property "onrelease" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: sentinel"
FAIL Navigator interface: attribute wakeLock assert_true: The prototype object must have a property "wakeLock" expected true got false
FAIL Navigator interface: navigator must inherit property "wakeLock" with the proper type assert_inherits: property "wakeLock" not found in prototype chain
